requirements.txt
# requirements.txt 파일의 내용

Flask==2.0.3
torch==1.9.1
torchvision==0.10.1
opencv-python==4.5.3.56
python-dotenv==0.19.0
requests==2.26.0
